% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circIMPACT_functions.R
\name{circ_contrib}
\alias{circ_contrib}
\title{Select circRNA markers
To establish a possible impact of circRNA in gene expression
we select a subset of circRNA defined as markers: 
only circRNA which expression can discretize two groups 
(DESeq adj. p-value<.01) have been selected.}
\usage{
circ_contrib(matrix_pc, K)
}
\arguments{
\item{dat}{circRNA data set}

\item{dds}{DeseqDataSet object filterd}

\item{sf}{sizefactor}

\item{p.cutoff}{select only circRNAs signficantly differentially expressed}

\item{lfc.cutoff}{log fold change cutoff to select circRNAs highly deregulated}

\item{k}{number of group in K-means algorithm}

\item{method.d}{method for calculate matrix distance}

\item{method.c}{method for clustering}
}
\value{
Table of top K circRNA for subsequent analysis
}
\description{
Select circRNA markers
To establish a possible impact of circRNA in gene expression
we select a subset of circRNA defined as markers: 
only circRNA which expression can discretize two groups 
(DESeq adj. p-value<.01) have been selected.
}
