% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circRNA_data.R
\docType{data}
\name{circularData}
\alias{circularData}
\title{circular RNA count matrix}
\format{
A matrix:
\describe{
  \item{row}{circRNA IDs}
  \item{column}{Samples}
}
}
\usage{
circularData
}
\description{
A data set with the row count matrix of circRNA across samples
}
\keyword{datasets}
