% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTarget_functions.R
\name{color_tile3}
\alias{color_tile3}
\title{Define formatter function for plot table}
\usage{
color_tile3(fun = "comma", digits = 3, palette = "PiYG", n = 10)
}
\description{
Define formatter function for plot table
}
