% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGAP_functions.R
\name{color_tile4}
\alias{color_tile4}
\title{Define formatter function for plot table}
\usage{
color_tile4(fun = "comma", digits = 0, palette = "YlGnBu", n = 9)
}
\description{
Define formatter function for plot table
}
