% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circIMPACT_functions.R
\name{contrib}
\alias{contrib}
\title{Define contribution selection function for ranking}
\usage{
contrib(ind.coord, comp.sdev, n.ind)
}
\description{
Define contribution selection function for ranking
}
