% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_data.R
\docType{data}
\name{count.matrix}
\alias{count.matrix}
\title{linear RNA count matrix}
\format{
A matrix:
\describe{
  \item{row}{mRNA IDs}
  \item{column}{Samples}
}
}
\usage{
count.matrix
}
\description{
A data set with the row count matrix of mRNA across samples
}
\keyword{datasets}
