% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIGAP_functions.R
\name{export_formattable}
\alias{export_formattable}
\title{Export a Formattable as PNG, PDF, or JPEG}
\usage{
export_formattable(
  f,
  file,
  width = "100\%",
  height = "100\%",
  background = "white",
  delay = 10
)
}
\arguments{
\item{f}{A formattable.}

\item{file}{Export path with extension .png, .pdf, or .jpeg.}

\item{width}{Width specification of the html widget being exported.}

\item{height}{Height specification of the html widget being exported.}

\item{background}{Background color specification.}

\item{delay}{Time to wait before taking webshot, in seconds.}
}
\description{
Export a Formattable as PNG, PDF, or JPEG
}
