% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circIMPACT_functions.R
\name{gene_class}
\alias{gene_class}
\title{Detect the most discriminant genes using circRNA-impact as stratificator of samples}
\usage{
gene_class(
  circ_idofinterest,
  circRNAs,
  linearRNAs,
  group,
  colData,
  covariates,
  th.corr
)
}
\arguments{
\item{circ_idofinterest}{circRNA-impact ID}

\item{circRNAs}{normalized read count of circRNAs}

\item{linearRNAs}{row read count of gene}

\item{colData}{coldata dataframe}

\item{covariates}{additinal covariates for batch effect}

\item{th.corr}{correlation thershold in clean collinear genes}
}
\value{
list of discriminating genes, n. genes selected, data.frame of ranked genes
}
\description{
Detect the most discriminant genes using circRNA-impact as stratificator of samples
}
