% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTarget_functions.R
\name{geneexpression}
\alias{geneexpression}
\title{Detect dergulated genes using circRNA-markers as stratificator of samples}
\usage{
geneexpression(
  circ_idofinterest,
  circRNAs,
  linearRNAs,
  group,
  colData,
  covariates,
  padj
)
}
\arguments{
\item{circ_idofinterest}{circRNA-marker ID}

\item{circRNAs}{normalized read count of circRNAs}

\item{linearRNAs}{row read count of gene}

\item{colData}{coldata dataframe}

\item{covariates}{additinal covariates for batch effect}

\item{meta}{meta file}
}
\value{
data.frame of gene deregulated per circRNA-markers
}
\description{
Detect dergulated genes using circRNA-markers as stratificator of samples
}
