% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTarget_functions.R
\name{get.color.hues}
\alias{get.color.hues}
\title{Define colors of condition}
\usage{
get.color.hues(meta)
}
\arguments{
\item{meta}{meta file}
}
\value{
data.frame of color hex and hue per condition
}
\description{
Define colors of condition
}
