% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circIMPACT_functions.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Define colors of condition}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of condion}
}
\value{
data.frame of color hues per condition
}
\description{
Define colors of condition
}
