% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTarget_functions.R
\name{marker.detection}
\alias{marker.detection}
\title{Detect circRNA markers
To establish a possible impact of circRNA in gene expression
we define two groups of samples defined by similarity in circRNAs expression.
K-means algorithm is used to define these two groups and then DESeq test (adj. p-value<.01) 
have been used to detect circRNA markers.}
\usage{
marker.detection(
  circ_id,
  circ.m,
  dds,
  sf,
  method.d,
  method.c,
  k,
  median = TRUE
)
}
\arguments{
\item{circ.m}{circRNA id}

\item{dds}{DeseqDataSet object}

\item{sf}{sizefactor}

\item{method.d}{method for calculate matrix distance}

\item{method.c}{method for clustering}

\item{k}{number of group in K-means algorithm}
}
\value{
data.frame of padj
}
\description{
Detect circRNA markers
To establish a possible impact of circRNA in gene expression
we define two groups of samples defined by similarity in circRNAs expression.
K-means algorithm is used to define these two groups and then DESeq test (adj. p-value<.01) 
have been used to detect circRNA markers.
}
