% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{nazero}
\alias{nazero}
\title{CircTarget}
\usage{
nazero(countmatrix)
}
\arguments{
\item{countmatrix}{}
}
\value{
list
}
\description{
define circRNA markers
}
\examples{
dataset <- matrix(sample(c(NA, 1:5), 25, replace = TRUE), 5)
df <- as.data.frame(dataset)
package::nazero(df)

}
