% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circTarget_functions.R
\name{stoplighttile}
\alias{stoplighttile}
\title{Define formatter function for plot table}
\usage{
stoplighttile(cut1 = 0.01, cut2 = 0.05, cut3 = 0.1, fun = "comma", digits = 4)
}
\description{
Define formatter function for plot table
}
