% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circIMPACT_functions.R
\name{geneexpression}
\alias{geneexpression}
\title{Detect dergulated genes using circRNA-markers as stratificator of samples}
\usage{
geneexpression(
  circ_idofinterest,
  circRNAs,
  linearRNAs,
  group,
  colData,
  covariates,
  padj,
  expr = FALSE
)
}
\arguments{
\item{circ_idofinterest}{circRNA-marker ID}

\item{circRNAs}{normalized read count of circRNAs}

\item{linearRNAs}{row read count of gene}

\item{colData}{coldata dataframe}

\item{covariates}{additinal covariates for batch effect}

\item{padj}{alpha thershold in deseq test}

\item{expr}{A TRUE or FALSE variable controlling whether the median expression should be used to define the two group comparison in DEG analysis.}
}
\value{
data.frame of gene deregulated per circRNA-markers
}
\description{
Detect dergulated genes using circRNA-markers as stratificator of samples
}
